refactor: extract browser-extension seek functions into lib module (Phase 2)
Extract 28 seek-related helper functions from content.ts into
lib/seek-extractor.ts using dependency-injection factory pattern.
Content.ts shrinks from 7229 to 6880 lines (-349).

Extracted functions: isSeekProfilePage, isSeekTalentSearchListPage,
getCurrentSeekMode, isSeekInlineProfileMode, isSeekProfileMode,
hasSeekProfileSnapshot, hasSeekListSnapshot, hasSeekTalentSearchSnapshot,
getSeekSnapshotCount, isSeekSnapshotReady, getSeekCandidateIdentity,
buildSeekProfileUrl, buildSeekNameSearchUrl, normalizeSeekLocationLabel,
restoreSeekSearchParams, getSeekRecommendedRequest,
getSeekTalentSearchRequest, getSeekProfileRequest, getSeekAutoSyncHelpers,
resolveSeekAutoSyncPageSize, resolveSeekAutoSyncPageWindow,
isSeekAutoSyncPageWindowReached, resolveSeekAutoSyncCurrentPageSelection,
getSeekRequestedPageSize, getSeekCurrentCandidateCount,
setSeekAutoSyncWindowAttributes, setSeekAutoSyncSelectionAttributes,
findSeekProfileTrigger.

All 28 functions verified against inline implementations — 8 mismatches
found and fixed (restoreSeekSearchParams storage key, normalizeSeekLocationLabel
rules, getSeekCandidateIdentity return shape, buildSeekProfileUrl URL pattern,
buildSeekNameSearchUrl signature, setSeek*Attributes missing persist call,
findSeekProfileTrigger backticks).
